requirements.txt
python-dxf==7.7.0